Refactor imports in IndexDetails, IndexSection, and IndicesOverview components for consistency
