+++ graylog2-web-interface/src/components/indices/IndexDetails.tsx
@@ -23,7 +23,9 @@ import { defaultOnError } from 'util/conditional/onError';
 import NumberUtils from 'util/NumberUtils';
 import { Col, Row, Button } from 'components/bootstrap';
 import { Spinner } from 'components/common';
-import { IndexRangeSummary, ShardMeter, ShardRoutingOverview } from 'components/indices';
+import IndexRangeSummary from 'components/indices/IndexRangeSummary';
+import ShardMeter from 'components/indices/ShardMeter';
+import ShardRoutingOverview from 'components/indices/ShardRoutingOverview';
 import type { IndexInfo } from 'stores/indices/IndicesStore';
 import type { IndexRange } from 'stores/indexers/IndexerOverviewStore';
 import { IndicesActions } from 'stores/indices/IndicesStore';
