+++ graylog2-web-interface/src/components/indices/IndexSection.tsx
@@ -20,7 +20,9 @@ import styled, { css } from 'styled-components';
 import type { IndexInfo } from 'stores/indices/IndicesStore';
 import { Row, Col } from 'components/bootstrap';
 import type { IndexSummary as IndexSummaryType } from 'stores/indexers/IndexerOverviewStore';
-import { ClosedIndexDetails, IndexDetails, IndexSummary } from 'components/indices';
+import ClosedIndexDetails from 'components/indices/ClosedIndexDetails';
+import IndexDetails from 'components/indices/IndexDetails';
+import IndexSummary from 'components/indices/IndexSummary';
 import NumberUtils from 'util/NumberUtils';
 
 const Index = ({
